@@ -1,12 +1,11 @@
-import {SensorComponent} from 'esphome-ts/dist';
-import {tap} from 'rxjs/operators';
-import {PlatformAccessory, Service as HAPService} from 'homebridge';
-import {Characteristic, Service} from '../index';
+import { tap } from 'rxjs/operators';
+import { Characteristic, Service } from '../index';
+import { SensorComponent } from 'esphome-ts';
+import { PlatformAccessory, Service as HAPService } from 'homebridge';
 
 const isTemperatureComponent = (unitOfMeasurement: unknown) => unitOfMeasurement === '°C' || unitOfMeasurement === '°F';
 
 export const sensorHelper = (component: SensorComponent, accessory: PlatformAccessory): boolean => {
-
     if (isTemperatureComponent(component.unitOfMeasurement)) {
         defaultSetup(component, accessory, Service.TemperatureSensor, Characteristic.CurrentTemperature);
         return true;
@@ -17,16 +16,20 @@ export const sensorHelper = (component: SensorComponent, accessory: PlatformAcce
     return false;
 };
 
-const defaultSetup = (component: SensorComponent,
+const defaultSetup = (
+    component: SensorComponent,
     accessory: PlatformAccessory,
     SelectedService: typeof Service.TemperatureSensor | typeof Service.HumiditySensor,
-    SelectedCharacteristic: typeof Characteristic.CurrentTemperature | typeof Characteristic.CurrentRelativeHumidity): void => {
-    let temperatureSensor: HAPService | undefined = accessory.services.find((service) => service.UUID === SelectedService.UUID);
+    SelectedCharacteristic: typeof Characteristic.CurrentTemperature | typeof Characteristic.CurrentRelativeHumidity,
+): void => {
+    let temperatureSensor: HAPService | undefined = accessory.services.find(
+        (service) => service.UUID === SelectedService.UUID,
+    );
     if (!temperatureSensor) {
         temperatureSensor = accessory.addService(new SelectedService(component.name, ''));
     }
 
-    component.state$.pipe(
-        tap(() => temperatureSensor?.getCharacteristic(SelectedCharacteristic)?.setValue(component.value!)),
-    ).subscribe();
+    component.state$
+        .pipe(tap(() => temperatureSensor?.getCharacteristic(SelectedCharacteristic)?.setValue(component.value!)))
+        .subscribe();
 };
